requirements.txt
Flask==3.0.3
gunicorn==22.0.0
SQLAlchemy==2.0.31
psycopg2-binary==2.9.9
python-dotenv==1.0.1
Werkzeug==3.0.3
requests==2.32.3
google-cloud-firestore==2.16.1
google-cloud-secret-manager==2.20.2
google-auth==2.47.0
